# bin/main/br/com/generated/api/FactRegistryApi.kt
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
*/
package br.com.generated.api

import br.com.generated.model.EntityDefinitionResponse
import br.com.generated.model.FactDefinitionResponse
import org.springframework.http.HttpStatus
import org.springframework.http.MediaType
import org.springframework.http.ResponseEntity

import org.springframework.web.bind.annotation.*
import org.springframework.validation.annotation.Validated
import org.springframework.web.context.request.NativeWebRequest
import org.springframework.beans.factory.annotation.Autowired

import jakarta.validation.constraints.DecimalMax
import jakarta.validation.constraints.DecimalMin
import jakarta.validation.constraints.Email
import jakarta.validation.constraints.Max
import jakarta.validation.constraints.Min
import jakarta.validation.constraints.NotNull
import jakarta.validation.constraints.Pattern
import jakarta.validation.constraints.Size
import jakarta.validation.Valid

import kotlin.collections.List
import kotlin.collections.Map

@RestController
@Validated
interface FactRegistryApi {


    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/v1/decision/entities"],
            produces = ["application/json"]
    )
    fun listEntities(): ResponseEntity<List<EntityDefinitionResponse>>


    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/v1/decision/facts"],
            produces = ["application/json"]
    )
    fun listFacts( @Valid @RequestParam(value = "entity", required = false) entity: kotlin.String?, @Valid @RequestParam(value = "enabled", required = false) enabled: kotlin.Boolean?): ResponseEntity<List<FactDefinitionResponse>>
}