# bin/main/br/com/generated/api/RuleConfigurationApi.kt
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
*/
package br.com.generated.api

import br.com.generated.model.ConfigurationVersionResponse
import br.com.generated.model.CreateRuleConfigurationRequest
import br.com.generated.model.ErrorResponse
import br.com.generated.model.RuleConfigurationResponse
import br.com.generated.model.UpdateRuleConfigurationRequest
import org.springframework.http.HttpStatus
import org.springframework.http.MediaType
import org.springframework.http.ResponseEntity

import org.springframework.web.bind.annotation.*
import org.springframework.validation.annotation.Validated
import org.springframework.web.context.request.NativeWebRequest
import org.springframework.beans.factory.annotation.Autowired

import jakarta.validation.constraints.DecimalMax
import jakarta.validation.constraints.DecimalMin
import jakarta.validation.constraints.Email
import jakarta.validation.constraints.Max
import jakarta.validation.constraints.Min
import jakarta.validation.constraints.NotNull
import jakarta.validation.constraints.Pattern
import jakarta.validation.constraints.Size
import jakarta.validation.Valid

import kotlin.collections.List
import kotlin.collections.Map

@RestController
@Validated
interface RuleConfigurationApi {


    @RequestMapping(
            method = [RequestMethod.POST],
            value = ["/v1/decision/rule-configurations/{id}/activate"],
            produces = ["application/json"]
    )
    fun activateConfiguration( @PathVariable("id") id: java.util.UUID): ResponseEntity<RuleConfigurationResponse>


    @RequestMapping(
            method = [RequestMethod.POST],
            value = ["/v1/decision/rules/{ruleCode}/configurations"],
            produces = ["application/json"],
            consumes = ["application/json"]
    )
    fun createRuleConfiguration( @PathVariable("ruleCode") ruleCode: kotlin.String, @Valid @RequestBody createRuleConfigurationRequest: CreateRuleConfigurationRequest): ResponseEntity<RuleConfigurationResponse>


    @RequestMapping(
            method = [RequestMethod.POST],
            value = ["/v1/decision/rule-configurations/{id}/deactivate"],
            produces = ["application/json"]
    )
    fun deactivateConfiguration( @PathVariable("id") id: java.util.UUID): ResponseEntity<RuleConfigurationResponse>


    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/v1/decision/rule-configurations/{id}"],
            produces = ["application/json"]
    )
    fun getConfigurationById( @PathVariable("id") id: java.util.UUID): ResponseEntity<RuleConfigurationResponse>


    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/v1/decision/rule-configurations/{id}/versions"],
            produces = ["application/json"]
    )
    fun getVersionHistory( @PathVariable("id") id: java.util.UUID): ResponseEntity<List<ConfigurationVersionResponse>>


    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/v1/decision/rules/{ruleCode}/configurations"],
            produces = ["application/json"]
    )
    fun listConfigurationsByRuleCode( @PathVariable("ruleCode") ruleCode: kotlin.String): ResponseEntity<List<RuleConfigurationResponse>>


    @RequestMapping(
            method = [RequestMethod.PUT],
            value = ["/v1/decision/rule-configurations/{id}"],
            produces = ["application/json"],
            consumes = ["application/json"]
    )
    fun updateRuleConfiguration( @PathVariable("id") id: java.util.UUID, @Valid @RequestBody updateRuleConfigurationRequest: UpdateRuleConfigurationRequest): ResponseEntity<RuleConfigurationResponse>
}