# bin/main/static/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Decision Engine — Rule Platform API
  description: |
    # Decision Engine — Plataforma de Regras de PLD

    API da plataforma de decisão para Prevenção à Lavagem de Dinheiro (PLD).
    Desacopla a lógica de **detecção** (Screening Context) da lógica de **decisão**
    (quando gerar alertas, com quais critérios, baseado em quais fatos).

    ## Arquitetura

    ```
    ┌─────────────────┐     DetectionEvent      ┌──────────────────┐    DecisionMadeEvent    ┌───────────────┐
    │  Screening      │ ──────────────────────▶ │  Decision        │ ─────────────────────▶ │  Alert        │
    │  Context (MF09) │                         │  Context         │                        │  Context      │
    │                 │                         │                  │                        │               │
    │  Detecta        │                         │  Avalia regras   │                        │  Cria alertas │
    │  keywords       │                         │  Decide ações    │                        │  com status   │
    └─────────────────┘                         └──────────────────┘                        └───────────────┘
    ```

    ## Fluxo de Decisão

    ```
    1. Transação PIX recebida
    2. KeywordScreeningService detecta termo restrito → publica DetectionEvent
    3. Decision Engine consome evento:
       a) Verifica idempotência (transactionId + ruleId)
       b) Carrega configuração ativa da regra
...
    5. Publica DecisionMadeEvent
...

...

...

...

...
servers:
...
paths:
...
  /v1/decision/facts:
    get:
      tags: [Fact Registry]
      summary: Listar fatos disponíveis
      description: |
        Retorna os fatos (Facts) disponíveis para uso em configurações de regras.
        Cada fact possui tipo, operadores suportados e entidade de origem.

        O analista consulta este catálogo para saber quais facts pode usar ao criar configurações.
      operationId: listFacts
      parameters:
        - name: entity
          in: query
          required: false
          description: "Filtrar por entidade de origem (ex.: Risk, Screening, Customer)"
          schema:
            type: string
          example: Risk
        - name: enabled
          in: query
          required: false
          description: Filtrar por status habilitado/desabilitado
          schema:
            type: boolean
          example: true
      responses:
        '200':
          description: Lista de definições de fatos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FactDefinitionResponse'
              examples:
                mvp:
                  summary: Fatos do MVP
                  value:
                    - id: "661e8400-e29b-41d4-a716-446655440001"
                      name: "keywordMatched"
                      displayName: "Keyword Matched"
                      entity: "Screening"
                      type: "BOOLEAN"
                      context: "SCREENING"
                      source: "Screening"
                      supportedOperators: ["EQUALS", "NOT_EQUALS"]
                      enabled: true
                    - id: "661e8400-e29b-41d4-a716-446655440002"
                      name: "customerRisk"
                      displayName: "Customer Risk"
                      entity: "Risk"
                      type: "ENUM"
                      context: "SCREENING"
                      source: "PLD"
                      supportedOperators: ["EQUALS", "NOT_EQUALS", "GREATER_THAN_OR_EQUAL"]
                      enabled: true
...
  /v1/decision/rules/{ruleCode}/configurations:
    post:
      tags: [Rule Configuration]
      summary: Criar configuração de regra
      description: |
        Cria uma nova Rule Configuration em estado **draft** (não ativa).
        O analista define quais condições (expressions) devem ser satisfeitas
        e quais ações executar quando todas forem verdadeiras.

        **Validações aplicadas:**
        - Cada `factName` deve existir no Fact Registry e estar habilitado
        - O `factName` deve estar nos `supportedFacts` da Rule Definition
        - O `operator` deve estar nos `supportedOperators` do Fact
        - O tipo do `expectedValue` deve ser compatível com o tipo do Fact
        - Máximo 10 expressions por configuração

        **Semântica AND implícita:** ALERT é gerado somente quando TODAS as condições
        são satisfeitas simultaneamente.
      operationId: createRuleConfiguration
      parameters:
        - name: ruleCode
          in: path
          required: true
          description: "Código da Rule Definition (ex.: KEYWORD_SCREENING)"
          schema:
            type: string
          example: KEYWORD_SCREENING
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRuleConfigurationRequest'
            examples:
              keywordAndRisk:
                summary: Keyword match + risco médio ou alto
                description: Gera alerta quando detecta keyword E cliente tem risco >= MR
                value:
                  expressions:
                    - type: "CONDITION"
                      factName: "keywordMatched"
                      operator: "EQUALS"
                      expectedValue: true
                    - type: "CONDITION"
                      factName: "customerRisk"
                      operator: "GREATER_THAN_OR_EQUAL"
                      expectedValue: "MR"
                  actions: ["GENERATE_ALERT"]
                  createdBy: "analyst@company.com"
              keywordOnly:
                summary: Apenas keyword match (qualquer risco)
                value:
                  expressions:
                    - type: "CONDITION"
                      factName: "keywordMatched"
                      operator: "EQUALS"
                      expectedValue: true
                  actions: ["GENERATE_ALERT"]
                  createdBy: "analyst@company.com"
      responses:
        '201':
          description: Configuração criada em estado draft (active=false, draft=true, version=1)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleConfigurationResponse'
        '400':
          description: Validação de campos obrigatórios falhou
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Rule Definition não encontrada para o ruleCode informado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: |
            Configuração inválida. Possíveis causas:
            - Fact não existe no catálogo
            - Fact está desabilitado
            - Fact não é suportado pela regra
            - Operador incompatível com o Fact
            - Tipo do expectedValue incompatível com o tipo do Fact
            - Mais de 10 expressions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                factNotFound:
                  summary: Fact inexistente
                  value:
                    timestamp: "2025-01-15T10:30:00Z"
                    status: 422
                    error: "Unprocessable Entity"
                    message: "Fato 'unknownFact' não encontrado no catálogo"
                    details: null

    get:
      tags: [Rule Configuration]
      summary: Listar configurações por ruleCode
      description: Retorna todas as configurações (draft e active) associadas a uma regra.
      operationId: listConfigurationsByRuleCode
      parameters:
        - name: ruleCode
          in: path
          required: true
          schema:
            type: string
          example: KEYWORD_SCREENING
      responses:
        '200':
          description: Lista de configurações (pode ser vazia se nenhuma configurada)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RuleConfigurationResponse'
  /v1/decision/rule-configurations/{id}:
    get:
      tags: [Rule Configuration]
      summary: Buscar configuração por ID
      operationId: getConfigurationById
      parameters:
        - name: id
          in: path
          required: true
          description: UUID da configuração
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Configuração encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleConfigurationResponse'
        '404':
          description: Configuração não encontrada

    put:
      tags: [Rule Configuration]
      summary: Atualizar configuração
      description: |
        Atualiza expressions e/ou actions de uma configuração existente.
        Cada update cria uma nova versão (monotonicamente crescente).
        Versões anteriores são preservadas no histórico para auditoria.
      operationId: updateRuleConfiguration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRuleConfigurationRequest'
      responses:
        '200':
          description: Configuração atualizada (nova versão criada)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleConfigurationResponse'
        '404':
          description: Configuração não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validação das novas expressions falhou
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/decision/rule-configurations/{id}/activate:
    post:
      tags: [Rule Configuration]
      summary: Ativar configuração
      description: |
        Ativa uma configuração para que passe a avaliar transações reais.

        **Pré-requisitos:**
        1. Deve existir ao menos um dry-run executado na versão atual
        2. Não pode existir outra configuração ativa para a mesma regra

        ```
        ┌────────┐     dry-run     ┌──────────┐    activate    ┌────────┐
        │ DRAFT  │ ──────────────▶ │ TESTADA  │ ─────────────▶ │ ACTIVE │
        │        │                 │          │                │        │
        └────────┘                 └──────────┘                └────────┘
        ```
      operationId: activateConfiguration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Configuração ativada (active=true, draft=false)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleConfigurationResponse'
        '404':
          description: Configuração não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Já existe configuração ativa para a mesma regra
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                conflict:
                  value:
                    timestamp: "2025-01-15T10:30:00Z"
                    status: 409
                    error: "Conflict"
                    message: "Já existe uma configuração ativa para a regra 'KEYWORD_SCREENING'"
                    details: null
        '422':
          description: Dry-run obrigatório antes da ativação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                noDryRun:
                  value:
                    timestamp: "2025-01-15T10:30:00Z"
                    status: 422
                    error: "Unprocessable Entity"
                    message: "Dry-run obrigatório antes da ativação"
                    details: null
  /v1/decision/rule-configurations/{id}/deactivate:
    post:
      tags: [Rule Configuration]
      summary: Desativar configuração
      description: |
        Desativa uma configuração. Transações não serão mais avaliadas por esta configuração.
        A configuração permanece no sistema para consulta e pode ser reativada após novo dry-run.
      operationId: deactivateConfiguration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Configuração desativada (active=false)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleConfigurationResponse'
        '404':
          description: Configuração não encontrada
  /v1/decision/rule-configurations/{id}/versions:
    get:
      tags: [Rule Configuration]
      summary: Histórico de versões
      description: |
        Retorna todas as versões de uma configuração em ordem cronológica.
        Cada update cria uma nova versão — versões anteriores são imutáveis.
      operationId: getVersionHistory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de versões (ordenada por version asc)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConfigurationVersionResponse'
        '404':
          description: Configuração não encontrada

  # ═══════════════════════════════════════════════════════════════════════════════
  # DRY-RUN — Simulação de configurações
  # ═══════════════════════════════════════════════════════════════════════════════
...
components:
  schemas:
...
    RuleContext:
      type: string
      description: |
        Contexto ao qual uma Rule Definition pertence.
        Identifica o domínio de avaliação da regra.
      enum: [SCREENING, TRANSACTION, CUSTOMER, ACCOUNT]
      example: SCREENING
...
    ComparisonOperator:
      type: string
      description: |
        Operador de comparação para expressões.
        **MVP suporta:** EQUALS, NOT_EQUALS, GREATER_THAN_OR_EQUAL.
        Outros operadores estão modelados para evolução futura.
      enum: [EQUALS, NOT_EQUALS, GREATER_THAN, GREATER_THAN_OR_EQUAL, LESS_THAN, LESS_THAN_OR_EQUAL, IN, NOT_IN, CONTAINS]
      example: EQUALS
    Action:
      type: string
      description: |
        Ação a ser executada quando a decisão é ALERT.
        - **GENERATE_ALERT**: Cria alerta no Alert Context
        - **IGNORE**: Nenhuma ação (apenas registro)
        - **REVIEW**: (futuro) Enviar para revisão
        - **BLOCK**: (futuro) Bloquear
      enum: [GENERATE_ALERT, IGNORE, REVIEW, BLOCK]
      example: GENERATE_ALERT
    FactType:
      type: string
      description: |
        Tipo do valor de um Fact.
        Determina quais operadores são válidos e como o valor é interpretado.
      enum: [BOOLEAN, ENUM, MONEY, STRING, NUMBER]
      example: BOOLEAN

    # ═══════════════════════════════════════════════════════════════════════════════
    # REQUEST DTOs
    # ═══════════════════════════════════════════════════════════════════════════════
    ConditionDto:
      type: object
      description: |
        Expressão atômica (condição) que compara um Fact contra um valor esperado.
        No MVP, `type` é sempre "CONDITION". Groups (AND/OR compostos) serão suportados em versões futuras.
      required: [type, factName, operator, expectedValue]
      properties:
        type:
          type: string
          description: Tipo da expressão. MVP suporta apenas "CONDITION".
          example: "CONDITION"
          enum: [CONDITION]
        factName:
          type: string
          description: |
            Nome do Fact a ser avaliado. Deve existir no Fact Registry e estar habilitado.
            Consulte `GET /v1/decision/facts` para ver fatos disponíveis.
          example: "keywordMatched"
        operator:
          $ref: '#/components/schemas/ComparisonOperator'
          description: |
            Operador de comparação. Deve estar nos `supportedOperators` do Fact.
            Ex.: keywordMatched suporta EQUALS e NOT_EQUALS.
            Ex.: customerRisk suporta EQUALS, NOT_EQUALS e GREATER_THAN_OR_EQUAL.
        expectedValue:
          description: |
            Valor esperado para a comparação. O tipo deve ser compatível com o `type` do Fact:
            - BOOLEAN → `true` ou `false`
            - ENUM → String com valor do enum (ex.: "MR", "AR", "BR")
            - NUMBER → Número (ex.: 1000.50)
            - STRING → String
            - MONEY → Objeto `{"amount": 1000.00, "currency": "BRL"}`
          oneOf:
            - type: boolean
            - type: string
            - type: number
            - type: object
          example: true
    CreateRuleConfigurationRequest:
      type: object
      description: Request para criar uma nova configuração de regra em estado draft.
      required: [expressions, actions, createdBy]
      properties:
        expressions:
          type: array
          description: |
            Lista de condições (semântica AND implícita).
            Todas devem ser satisfeitas para que a decisão seja ALERT.
            Mínimo 1, máximo 10.
          minItems: 1
          maxItems: 10
          items:
            $ref: '#/components/schemas/ConditionDto'
        actions:
          type: array
          description: |
            Ações a executar quando TODAS as condições forem satisfeitas.
            No MVP: GENERATE_ALERT (cria alerta) ou IGNORE.
          minItems: 1
          items:
            $ref: '#/components/schemas/Action'
        createdBy:
          type: string
          description: Identificação de quem está criando (email ou username)
          maxLength: 100
          example: "analyst@company.com"
    UpdateRuleConfigurationRequest:
      type: object
      description: |
        Request para atualizar uma configuração existente.
        Cada update cria uma nova versão no histórico.
      required: [expressions, actions, updatedBy]
      properties:
        expressions:
          type: array
          description: Nova lista de condições (substitui as anteriores)
          minItems: 1
          maxItems: 10
          items:
            $ref: '#/components/schemas/ConditionDto'
        actions:
          type: array
          description: Nova lista de ações
          minItems: 1
          items:
            $ref: '#/components/schemas/Action'
        updatedBy:
          type: string
          description: Identificação de quem está atualizando
          maxLength: 100
          example: "analyst@company.com"
...
    FactDefinitionResponse:
      type: object
      description: |
        Definição de um Fact no catálogo.
        Descreve um dado contextual que pode ser usado em expressões de regras.
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único do fact
        name:
          type: string
          description: Nome técnico do fact (usado em expressions)
          example: "customerRisk"
        displayName:
          type: string
          description: Nome amigável para exibição na UI
          example: "Customer Risk"
        entity:
          type: string
          description: Entidade de negócio à qual o fact pertence
          example: "Risk"
        type:
          $ref: '#/components/schemas/FactType'
          description: |
            Tipo do valor. Determina qual tipo de `expectedValue` usar nas expressions:
            - BOOLEAN → true/false
            - ENUM → String (ex.: "BR", "MR", "AR")
            - NUMBER → Numérico
            - STRING → Texto livre
            - MONEY → Objeto {amount, currency}
        context:
          $ref: '#/components/schemas/RuleContext'
          description: Contexto onde o fact é utilizado
        source:
          type: string
          description: Bounded context de origem dos dados
          example: "PLD"
        supportedOperators:
          type: array
          description: |
            Operadores de comparação que este fact suporta.
            O analista só pode usar estes operadores ao criar expressões com este fact.
          items:
            $ref: '#/components/schemas/ComparisonOperator'
          example: ["EQUALS", "NOT_EQUALS", "GREATER_THAN_OR_EQUAL"]
        enabled:
          type: boolean
          description: Se o fact está habilitado para uso em novas configurações
          example: true
...
    RuleConfigurationResponse:
      type: object
      description: |
        Configuração de regra criada/gerenciada pelo analista.
        Contém as expressões de decisão e ações a serem executadas.
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único da configuração
        ruleCode:
          type: string
          description: Código da regra à qual esta configuração pertence
          example: "KEYWORD_SCREENING"
        version:
          type: integer
          description: Versão atual da configuração (incrementa a cada update)
          example: 1
        active:
          type: boolean
          description: |
            Se a configuração está ativa (avaliando transações em produção).
            Apenas uma configuração pode estar ativa por regra.
          example: false
        draft:
          type: boolean
          description: |
            Se a configuração está em rascunho (nunca foi ativada).
            draft=true e active=false → configuração recém-criada.
          example: true
        expressions:
          type: array
          description: Lista de condições (AND implícito entre todas)
          items:
            $ref: '#/components/schemas/ConditionDto'
        actions:
          type: array
          description: Ações a executar quando todas condições forem satisfeitas
          items:
            $ref: '#/components/schemas/Action'
        createdBy:
          type: string
          description: Quem criou esta configuração
          example: "analyst@company.com"
        createdAt:
          type: string
          format: date-time
          description: Data de criação
          example: "2025-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Data da última atualização (update, activate ou deactivate)
          example: "2025-01-15T10:30:00Z"
    ConfigurationVersionResponse:
      type: object
      description: |
        Snapshot de uma versão específica da configuração.
        Imutável após criação — representa o estado exato naquele ponto no tempo.
      properties:
        version:
          type: integer
          description: Número da versão (monotonicamente crescente, começa em 1)
          example: 1
        expressions:
          type: array
          description: Expressões vigentes nesta versão
          items:
            $ref: '#/components/schemas/ConditionDto'
        actions:
          type: array
          description: Ações configuradas nesta versão
          items:
            $ref: '#/components/schemas/Action'
        active:
          type: boolean
          description: Se esta versão estava ativa no momento da criação
        createdBy:
          type: string
          description: Quem criou esta versão
          example: "analyst@company.com"
        createdAt:
          type: string
          format: date-time
          description: Quando esta versão foi criada
          example: "2025-01-15T10:30:00Z"
...
    ErrorResponse:
      type: object
      description: |
        Formato padronizado de erro para todos os endpoints.
        Inclui timestamp, status HTTP, código de erro e mensagem descritiva.
      properties:
        timestamp:
          type: string
          format: date-time
          description: Momento do erro (ISO-8601)
          example: "2025-01-15T10:30:00Z"
        status:
          type: integer
          description: Código HTTP do erro
          example: 422
        error:
          type: string
          description: Descrição curta do tipo de erro HTTP
          example: "Unprocessable Entity"
        message:
          type: string
          description: Mensagem detalhada do erro (legível)
          example: "Fato 'unknownFact' não encontrado no catálogo"
        details:
          type: string
          description: "Detalhes adicionais (ex.: campos com erro de validação)"
          nullable: true
          example: "expressions[0].factName: não pode ser vazio; createdBy: é obrigatório"